What xlC really wanted to tell me was to use int instead of short...

+++ devel/bmake/files/lst.lib/lstInt.h
@@ -1,4 +1,4 @@
-/*	$NetBSD: lstInt.h,v 1.2 2007/07/31 13:21:46 joerg Exp $	*/
+/*	$NetBSD: lstInt.h,v 1.3 2007/07/31 13:28:57 joerg Exp $	*/
 
 /*
  * Copyright (c) 1988, 1989, 1990, 1993
@@ -47,7 +47,7 @@
 typedef struct ListNode {
 	struct ListNode	*prevPtr;   /* previous element in list */
 	struct ListNode	*nextPtr;   /* next in list */
-	unsigned short	 useCount:8, /* Count of functions using the node.
+	unsigned int	 useCount:8, /* Count of functions using the node.
 				     * node may not be deleted until count
 				     * goes to 0 */
  	    	    	flags:8;    /* Node status flags */
